Move halt pipeline signal to Payload
